+++ src/main/java/factorymethod/date/Date.java
@@ -1,4 +1,4 @@
-package factorymethod;
+package factorymethod.date;
 
 /**
  * Дата.
